# Improve speed of _freeze_time.start()