I upgraded some tests in my test suite to stop using `mock.patch` on datetime functions to using `freezegun.freeze_time`. I noticed they got a bit slower - this is not so surprising since `freeze_time` is more thorough. However it lead me to find some easy improvements to the inner loop in `start`, when all the modules are scanned, using the fantastic [line_profiler](https://github.com/rkern/line_profiler).

The changes are:

1. Making `ignore` a tuple always
2. Using `elif` for case fallthrough (this doesn't speed things up that much but it simplifies the jump statements the cpython VM uses)
3. Using tuples for `if x in list` - this is normally peephole-optimized into tuples by the CPython VM, but I don't know about any other python implementations such as pypy. When comparing to a fixed set, tuples are faster because they can be stored in the code object's co_consts.
4. Changing the attribute value comparison to use a local tuple `real_things` for the initial check. This avoids stepping through multiple `x is y` comparisons. The hit ratio for the `not in real_things` statement is around 99.86%. This is where most of the savings come from.
5. Localize `self.undo_changes.append`, to avoid the attribute lookup cost to call them.

This reduced the time spent in `start()` by nearly 50%, from 14.5s to 10.1s, when running `kernprof` on a unit test file that includes 5 `freeze_time` invocations. Profiling:

Before:
```
Timer unit: 1e-06 s

Total time: 14.5213 s
File: /path/to/freezegun/api.py
Function: start at line 265

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   265                                               def start(self):
   266                                                   # Change the modules
   267         5           60     12.0      0.0          datetime.datetime = FakeDatetime
   268         5           32      6.4      0.0          datetime.date = FakeDate
   269         5           54     10.8      0.0          fake_time = FakeTime(self.time_to_freeze, time.time)
   270         5           35      7.0      0.0          fake_localtime = FakeLocalTime(self.time_to_freeze)
   271         5           37      7.4      0.0          fake_gmtime = FakeGMTTime(self.time_to_freeze, time.gmtime)
   272         5           32      6.4      0.0          fake_strftime = FakeStrfTime(self.time_to_freeze)
   273         5           17      3.4      0.0          time.time = fake_time
   274         5           22      4.4      0.0          time.localtime = fake_localtime
   275         5           18      3.6      0.0          time.gmtime = fake_gmtime
   276         5           19      3.8      0.0          time.strftime = fake_strftime
   277                                           
   278         5           26      5.2      0.0          copyreg.dispatch_table[real_datetime] = pickle_fake_datetime
   279         5           17      3.4      0.0          copyreg.dispatch_table[real_date] = pickle_fake_date
   280                                           
   281                                                   # Change any place where the module had already been imported
   282     20778       108283      5.2      0.7          for mod_name, module in list(sys.modules.items()):
   283     20773        73788      3.6      0.5              if mod_name is None or module is None:
   284      9752        34064      3.5      0.2                  continue
   285     11021       114730     10.4      0.8              if mod_name.startswith(tuple(self.ignore)):
   286        50          168      3.4      0.0                  continue
   287     10971        63794      5.8      0.4              if (not hasattr(module, "__name__")
   288     10971        40310      3.7      0.3                  or module.__name__ in ['datetime', 'time']):
   289        10           30      3.0      0.0                  continue
   290    301500      1329914      4.4      9.2              for module_attribute in dir(module):
   291    290539      1017877      3.5      7.0                  if module_attribute in ['real_date', 'real_datetime',
   292    290539      1067735      3.7      7.4                      'real_time', 'real_localtime', 'real_gmtime', 'real_strftime']:
   293        65         1281     19.7      0.0                      continue
   294    290474      1020950      3.5      7.0                  try:
   295    290474      2121421      7.3     14.6                      attribute_value = getattr(module, module_attribute)
   296       208         1882      9.0      0.0                  except (ImportError, AttributeError):
   297                                                               # For certain libraries, this can result in ImportError(_winreg) or AttributeError (celery)
   298       208          712      3.4      0.0                      continue
   299    290266      1033340      3.6      7.1                  try:
   300    290266      1225441      4.2      8.4                      if attribute_value is real_datetime:
   301       255         2247      8.8      0.0                          setattr(module, module_attribute, FakeDatetime)
   302       255         2321      9.1      0.0                          self.undo_changes.append((module, module_attribute, real_datetime))
   303    290011      1051108      3.6      7.2                      elif attribute_value is real_date:
   304        70          608      8.7      0.0                          setattr(module, module_attribute, FakeDate)
   305        70          537      7.7      0.0                          self.undo_changes.append((module, module_attribute, real_date))
   306    289941      1051316      3.6      7.2                      elif attribute_value is real_time:
   307        75          522      7.0      0.0                          setattr(module, module_attribute, fake_time)
   308        75          640      8.5      0.0                          self.undo_changes.append((module, module_attribute, real_time))
   309    289866      1049201      3.6      7.2                      elif attribute_value is real_localtime:
   310        10           38      3.8      0.0                          setattr(module, module_attribute, fake_localtime)
   311        10           45      4.5      0.0                          self.undo_changes.append((module, module_attribute, real_localtime))
   312    289856      1036639      3.6      7.1                      elif attribute_value is real_gmtime:
   313                                                                   setattr(module, module_attribute, fake_gmtime)
   314                                                                   self.undo_changes.append((module, module_attribute, real_gmtime))
   315    289856      1069604      3.7      7.4                      elif attribute_value is real_strftime:
   316                                                                   setattr(module, module_attribute, fake_strftime)
   317                                                                   self.undo_changes.append((module, module_attribute, real_strftime))
   318                                                           except:
   319                                                               # If it's not possible to compare the value to real_XXX (e.g. hiredis.version)
   320                                                               pass
   321                                           
   322         5           64     12.8      0.0          datetime.datetime.times_to_freeze.append(self.time_to_freeze)
   323         5           27      5.4      0.0          datetime.datetime.tz_offsets.append(self.tz_offset)
   324                                           
   325                                                   # Since datetime.datetime has already been mocked, just use that for
   326                                                   # calculating the date
   327         5          309     61.8      0.0          datetime.date.dates_to_freeze.append(datetime.datetime.now().date())
```

After:
```
Timer unit: 1e-06 s

Total time: 10.161 s
File: /path/to/freezegun/api.py
Function: start at line 265

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   265                                               def start(self):
   266                                                   # Change the modules
   267         5           35      7.0      0.0          datetime.datetime = FakeDatetime
   268         5           26      5.2      0.0          datetime.date = FakeDate
   269         5           45      9.0      0.0          fake_time = FakeTime(self.time_to_freeze, time.time)
   270         5           31      6.2      0.0          fake_localtime = FakeLocalTime(self.time_to_freeze)
   271         5           35      7.0      0.0          fake_gmtime = FakeGMTTime(self.time_to_freeze, time.gmtime)
   272         5           28      5.6      0.0          fake_strftime = FakeStrfTime(self.time_to_freeze)
   273         5           15      3.0      0.0          time.time = fake_time
   274         5           20      4.0      0.0          time.localtime = fake_localtime
   275         5           15      3.0      0.0          time.gmtime = fake_gmtime
   276         5           20      4.0      0.0          time.strftime = fake_strftime
   277                                           
   278         5           24      4.8      0.0          copyreg.dispatch_table[real_datetime] = pickle_fake_datetime
   279         5           19      3.8      0.0          copyreg.dispatch_table[real_date] = pickle_fake_date
   280                                           
   281                                                   # Change any place where the module had already been imported
   282                                                   real_things = (
   283         5           16      3.2      0.0              real_time,
   284         5           16      3.2      0.0              real_localtime,
   285         5           14      2.8      0.0              real_gmtime,
   286         5           16      3.2      0.0              real_strftime,
   287         5           15      3.0      0.0              real_date,
   288         5           15      3.0      0.0              real_datetime
   289                                                   )
   290         5           18      3.6      0.0          add_change = self.undo_changes.append
   291                                           
   292     20788       192076      9.2      1.9          for mod_name, module in list(sys.modules.items()):
   293     20783        74249      3.6      0.7              if mod_name is None or module is None:
   294      9750        34634      3.6      0.3                  continue
   295     11033        78072      7.1      0.8              elif mod_name.startswith(self.ignore):
   296        50          162      3.2      0.0                  continue
   297     10983        56790      5.2      0.6              elif (not hasattr(module, "__name__")
   298     10983        41133      3.7      0.4                  or module.__name__ in ('datetime', 'time')):
   299        10           31      3.1      0.0                  continue
   300    301820      1312355      4.3     12.9              for module_attribute in dir(module):
   301    290847      1033415      3.6     10.2                  if module_attribute in ('real_date', 'real_datetime',
   302    290847      1066544      3.7     10.5                      'real_time', 'real_localtime', 'real_gmtime', 'real_strftime'):
   303        65          204      3.1      0.0                      continue
   304    290782      1030523      3.5     10.1                  try:
   305    290782      1749226      6.0     17.2                      attribute_value = getattr(module, module_attribute)
   306       208         2125     10.2      0.0                  except (ImportError, AttributeError):
   307                                                               # For certain libraries, this can result in ImportError(_winreg) or AttributeError (celery)
   308       208         1709      8.2      0.0                      continue
   309    290574      1049822      3.6     10.3                  try:
   310    290574      1355271      4.7     13.3                      if attribute_value not in real_things:
   311    290124      1073289      3.7     10.6                          continue
   312                                           
   313       415         1427      3.4      0.0                      if attribute_value is real_datetime:
   314       255         1542      6.0      0.0                          setattr(module, module_attribute, FakeDatetime)
   315       255         1272      5.0      0.0                          add_change((module, module_attribute, real_datetime))
   316       160          708      4.4      0.0                      elif attribute_value is real_date:
   317        70          495      7.1      0.0                          setattr(module, module_attribute, FakeDate)
   318        70          453      6.5      0.0                          add_change((module, module_attribute, real_date))
   319        90          376      4.2      0.0                      elif attribute_value is real_time:
   320        75         1387     18.5      0.0                          setattr(module, module_attribute, fake_time)
   321        75          432      5.8      0.0                          add_change((module, module_attribute, real_time))
   322        15           53      3.5      0.0                      elif attribute_value is real_localtime:
   323        10           38      3.8      0.0                          setattr(module, module_attribute, fake_localtime)
   324        10           36      3.6      0.0                          add_change((module, module_attribute, real_localtime))
   325         5           18      3.6      0.0                      elif attribute_value is real_gmtime:
   326                                                                   setattr(module, module_attribute, fake_gmtime)
   327                                                                   add_change((module, module_attribute, real_gmtime))
   328         5           18      3.6      0.0                      elif attribute_value is real_strftime:
   329                                                                   setattr(module, module_attribute, fake_strftime)
   330                                                                   add_change((module, module_attribute, real_strftime))
   331        35          106      3.0      0.0                  except:
   332                                                               # If it's not possible to compare the value to real_XXX (e.g. hiredis.version)
   333        35          185      5.3      0.0                      pass
   334                                           
   335         5           57     11.4      0.0          datetime.datetime.times_to_freeze.append(self.time_to_freeze)
   336         5           26      5.2      0.0          datetime.datetime.tz_offsets.append(self.tz_offset)
   337                                           
   338                                                   # Since datetime.datetime has already been mocked, just use that for
   339                                                   # calculating the date
   340         5          299     59.8      0.0          datetime.date.dates_to_freeze.append(datetime.datetime.now().date())
```