@spulec 2015-07-23
This is great. Thanks for the help!